Moved project files to the project dir
